integrating api interfaces
